+++ packages/react-native-gesture-handler/src/v3/hooks/useGestureCallbacks.ts
@@ -1,20 +1,18 @@
-import { useGestureStateChangeEvent } from './events/useGestureStateChangeEvent';
-import { useGestureHandlerEvent } from './events/useGestureHandlerEvent';
-import { useTouchEvent } from './events/useTouchEvent';
+import { useGestureStateChangeEvent } from './callbacks/js/useGestureStateChangeEvent';
+import { useGestureUpdateEvent } from './callbacks/js/useGestureUpdateEvent';
+import { useGestureTouchEvent } from './callbacks/js/useGestureTouchEvent';
 import { AnimatedEvent } from '../types';
 import { checkMappingForChangeProperties, isAnimatedEvent } from './utils';
 
-export function useGestureEvent(handlerTag: number, config: any) {
+export function useGestureCallbacks(handlerTag: number, config: any) {
   const onGestureHandlerStateChange = useGestureStateChangeEvent(
     handlerTag,
     config
   );
-  const onGestureHandlerEvent = useGestureHandlerEvent(handlerTag, config);
 
-  const onGestureHandlerTouchEvent = useTouchEvent(handlerTag, config);
-
+  const onGestureHandlerEvent = useGestureUpdateEvent(handlerTag, config);
+  const onGestureHandlerTouchEvent = useGestureTouchEvent(handlerTag, config);
   let onGestureHandlerAnimatedEvent: AnimatedEvent | undefined;
-
   if (isAnimatedEvent(config.onUpdate)) {
     for (const mapping of config.onUpdate._argMapping) {
       checkMappingForChangeProperties(mapping);
